+++ package.json
@@ -11,6 +11,7 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-helmet": "^6.1.0",
+    "react-router-dom": "^6.2.1",
     "react-spring": "^9.4.2",
     "react-transition-group": "^4.4.2",
     "react-use": "^17.3.2"
